New thread pool implementation

+++ src/options.cpp
@@ -302,7 +302,7 @@ options_t::options_t()
 #else
   alloc_chunkwise(ALLOC_SPARSE),
 #endif
-  num_procs((int)std::min(4U, std::thread::hardware_concurrency()))
+  num_procs(std::min(4U, std::thread::hardware_concurrency()))
 {
     if (num_procs < 1) {
         log_warn("Unable to detect number of hardware threads supported!"
